nav bar update and edit profile
hey boys, I have the nav bars working well and I thought I had the profile picture working well, but I am having troubles getting it to pull the file path back from the database, so when you re load the screen or change pages the  profile picture goes away, but the path to it stays in the database which is weird. Ill do more work on it soon.