added profile picture to the navbar, also added icon to update profile picture
the profile pictures in the nav bar need a little work with the styling but they are atless all working now. I also added a new icon to the profile page, the user can click on it and it will allow them to update their profile page. In the future i would like that button to just go straight to the files folder.