Worked on the admin Page for Tickets
Created a new php file for ticket requests for the admins to view. As of right now, I only have drawing data from DB into a table whilst having 2 separate forms to (soon) add in the data from the DDL. This will then be able to accept their change request (ticket) or discard it.